requirements.txt
# Web 框架
Flask==3.0.0

# 数据处理
pandas==2.2.3
numpy>=1.22.4

# HTTP 请求
requests==2.32.3
beautifulsoup4==4.12.3

# 环境变量管理
python-dotenv==1.0.1

# 日期时间处理
python-dateutil>=2.8.2
pytz>=2020.1

# 其他依赖
certifi>=2017.4.17
charset-normalizer>=2.0.0
idna>=2.5
urllib3>=1.21.1
